Work on the lenses.

+++ storage/framework/messages/0296e3581f5421441e00c3c26eb3dee21b80bce4.php.php
@@ -33,7 +33,7 @@ class LensController extends Controller
      */
     public function __construct()
     {
-        $this->middleware(['auth', 'verified', 'clearance'])->except(['show']);
+        $this->middleware(['auth', 'verified'])->except(['show']);
     }
 
     /**
@@ -43,29 +43,9 @@ class LensController extends Controller
      */
     public function index()
     {
-        // LENSES
-        // TODO: Test index
-        // TODO: Test creating a new lens (only authenticated!)
-        // TODO: Add flash_messages when lens is deleted or updated (see store)
+        $lenses = auth()->user()->lenses()->get();
 
-        // AUTHENTICATION
-        // TODO: Register page
-        //      After registering, nothing seems to happen
-        // TODO: If the country of residence is not yet set,
-        //         show a flash message to do so...
-        // TODO: Write user settings page and table for the DeepskyLog information
-        // TODO: Update admin page for the users, add extra information,
-        //        move operations in two different colums, use icons for operations
-        // TODO: Use authentication on the pages and in the layout.
-        // TODO: Write script to convert old observers table of DeepskyLog to laravel
-        // TODO: Write script to convert old lenses table of DeepskyLog to laravel
-        // TODO: Page to change observer settings
-        // TODO: Page to view observer
-        // TODO: Clean up source code
-
-        $lens = auth()->user()->lenses();
-
-        return view('layout.lens.view', compact($lens));
+        return view('layout.lens.view')->with('lenses', $lenses);
     }
 
     /**
@@ -89,19 +69,19 @@ class LensController extends Controller
      */
     public function store(Request $request)
     {
+        $request['observer_id'] = auth()->id();
+
         $validated = request()->validate(
             [
                 'observer_id' => 'required',
-                'name' => ['required', 'min:5'],
-                'factor' => 'required',
+                'name' => ['required', 'min:6'],
+                'factor' => ['required', 'numeric', 'min:0', 'max:10'],
             ]
         );
 
-        $validated['observer_id'] = auth()->id();
-
         Lens::create($validated);
 
-        flash()->success(_i('Lens "%s" created'), $request->name);
+        flash()->success(_i('Lens "%s" created', $request->name));
 
         // View the page with all lenses for the user
         return redirect('/lens');
@@ -116,9 +96,7 @@ class LensController extends Controller
      */
     public function show(Lens $lens)
     {
-        $this->authorize('view', $lens);
-
-        // TO WRITE
+        return view('layout.lens.show', ['lens' => $lens]);
     }
 
     /**
@@ -130,7 +108,7 @@ class LensController extends Controller
      */
     public function edit(Lens $lens)
     {
-        $this->authorize('edit', $lens);
+        $this->authorize('update', $lens);
 
         return view('layout.lens.create', ['lens' => $lens, 'update' => true]);
     }
@@ -145,27 +123,36 @@ class LensController extends Controller
      */
     public function update(Request $request, Lens $lens)
     {
+        $this->authorize('update', $lens);
+
+        $request['observer_id'] = $lens->observer_id;
+
         // If the factor is set, the name should also be set in the form.
         if ($request->has('factor')) {
             $validated = request()->validate(
                 [
                     'observer_id' => 'required',
-                    'name' => ['required', 'min:5'],
-                    'factor' => 'required',
-                ]
+                    'name' => ['required', 'min:6'],
+                    'factor' => ['required', 'numeric', 'min:0', 'max:10'],
+                    ]
             );
 
             $lens->update(['factor' => $request->get('factor')]);
             $lens->update(['name' => $request->get('name')]);
+
+            flash()->warning(_i('Lens "%s" updated', $lens->name));
         } else {
             // This is only reached when clicking the active checkbox in the
             // lens overview.
             if ($request->has('active')) {
                 $lens->active();
+                flash()->warning(_i('Lens "%s" is active', $lens->name));
             } else {
                 $lens->inactive();
+                flash()->warning(_i('Lens "%s" is not longer active', $lens->name));
             }
         }
+
 
         return redirect('/lens');
     }
@@ -179,10 +166,12 @@ class LensController extends Controller
      */
     public function destroy(Lens $lens)
     {
+        $this->authorize('update', $lens);
+
         $lens->delete();
 
-        flash()->warning(_i('Lens "%s" deleted'), $lens->name);
+        flash()->error(_i('Lens "%s" deleted', $lens->name));
 
-        return view('layout.lens.view');
+        return redirect('/lens');
     }
 }
